[CHORE] 코드 정리
