+++ src/helpers/IndexPage.js
@@ -1,0 +1,17 @@
+import ErrorPage from "../pages/error/ErrorPage";
+import Game from "../pages/game/Game";
+import Intro from "../pages/intro/Intro";
+import KakaoSign from "../pages/intro/kakao/KakaoSign";
+import SetUserInfo from "../pages/intro/kakao/SetUserInfo";
+import Loading from "../pages/loading/Loading";
+import Lobby from "../pages/lobby/Lobby";
+
+export const PAGE = {
+  Intro: <Intro />,
+  ErrorPage: <ErrorPage />,
+  Loading: <Loading />,
+  Lobby: <Lobby />,
+  Game: <Game />,
+  KakaoSign: <KakaoSign />,
+  SetUserInfo: <SetUserInfo />,
+};
