+++ src/helpers/animation.js
@@ -1,0 +1,1 @@
+export const ANIMATION = {};
