+++ src/pages/game/logic/OtherTurn.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { useSelector } from 'react-redux';
 import styled from 'styled-components';
-import { ICON } from '../../Icons';
+import { ICON } from '../../../helpers/Icons';
 
 const OtherTurn = () => {
   const {gameInfo}=useSelector(state=>state.gameSlice)
